Improving email a little

+++ programs/email/init-writing.sh
@@ -9,11 +9,20 @@ else
 	source ./env.sh
 fi
 
-CUSTOMIZED=./custom/template-org.html
-if [ -f $CUSTOMIZED ]; then
-	cp -f $CUSTOMIZED $FILE_ORG
-else
-	cp -f ./template-org.html $FILE_ORG
+echo
+echo "Continue where you were? (y/n, default y)"
+read -p "> " CONTINUE
+
+if [ "$CONTINUE" == "n" ]; then
+	# Backup previous email, just to be sure
+	cp -f $FILE_ORG $FILE_ORG.bak
+	# Prepare next email
+	CUSTOMIZED=./custom/template-org.html
+	if [ -f $CUSTOMIZED ]; then
+		cp -f $CUSTOMIZED $FILE_ORG
+	else
+		cp -f ./template-org.html $FILE_ORG
+	fi
 fi
 
 # Open temp file in web browser
